Fixed testcase -3

+++ Statistics.Test/StatsUnitTest.cs
@@ -12,8 +12,8 @@ namespace Statistics.Test
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<float>{(float)1.5, (float)8.9, (float)3.2, (float)4.5 });
-            float epsilon = 0.001F;
+                new List<double>{(double)1.5, (double)8.9, (double)3.2, (double)4.5 });
+            double epsilon = 0.001F;
             Assert.True(Math.Abs(computedStats.average - 4.525) <= epsilon);
             Assert.True(Math.Abs(computedStats.max - 8.9) <= epsilon);
             Assert.True(Math.Abs(computedStats.min - 1.5) <= epsilon);
@@ -24,14 +24,14 @@ namespace Statistics.Test
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<float>{});
+                new List<double>{});
             // All fields of computedStats (average, max, min) must be
             // Double.NaN (not-a-number), as described in
             // https://docs.microsoft.com/en-us/dotnet/api/system.double.nan?view=netcore-3.1
             // Specify the Assert statements here
-            Assert.True(computedStats.average.Equals(Double.NaN));
-            Assert.True(computedStats.max.Equals(Double.NaN));
-            Assert.True(computedStats.min.Equals(Double.NaN));
+            Assert.True(computedStats.average.Equals(double.NaN));
+            Assert.True(computedStats.max.Equals(double.NaN));
+            Assert.True(computedStats.min.Equals(double.NaN));
         }
         
         [Fact]
@@ -39,8 +39,8 @@ namespace Statistics.Test
         {
             var statsComputer = new StatsComputer();
             var computedStats = statsComputer.CalculateStatistics(
-                new List<float> {(float) Double.NaN, (float) 1.5, (float) 8.9 });
-            float epsilon = 0.001F;
+                new List<double> {(double) double.NaN, (double) 1.5, (double) 8.9 });
+            double epsilon = 0.001F;
             Assert.True(Math.Abs(computedStats.average - 5.2) <= epsilon);
             Assert.True(Math.Abs(computedStats.max - 8.9) <= epsilon);
             Assert.True(Math.Abs(computedStats.min - 1.5) <= epsilon);
